split agents
