+++ agent/__init__.py
@@ -1,2 +1,5 @@
-from .agent import DQNAgent, StrongAgent, RandomAgent
+from .dqn_agent import DQNAgent
+from .imitation_agent import ImitationAgent
+from .random_agent import RandomAgent
+from .strong_agent import StrongAgent
 from .replay_buffer import ReplayBuffer
