+++ agent/dqn_agent.py
@@ -2,8 +2,6 @@ import torch
 from torch import nn
 import torch_geometric
 import numpy as np
-from torch_geometric.data import Batch
-import ecole
 
 
 class BipartiteGraphConvolution(torch_geometric.nn.MessagePassing):
@@ -178,7 +176,7 @@ class DQNAgent:
         self.net = BipartiteGCN(device=device, var_nfeats=24, n_heads=50)
         self.opt = torch.optim.Adam(self.net.parameters())
 
-    def act(self, obs, action_set, deterministic):
+    def act(self, obs, action_set, deterministic, env):
         if deterministic:
             mask = np.ones(len(self.net.heads))
         else:
@@ -221,25 +219,3 @@ class DQNAgent:
 
     def eval(self):
         self.net.eval()
-
-
-class StrongAgent:
-    def __init__(self, env):
-        self.strong_branching_function = ecole.observation.StrongBranchingScores()
-        self.env = env
-
-    def before_reset(self, model):
-        self.strong_branching_function.before_reset(model)
-
-    def act(self, obs, action_set, deterministic):
-        scores = self.strong_branching_function.extract(self.env.model, False)[action_set]
-        return action_set[np.argmax(scores)], None
-
-    def eval(self):
-        pass
-
-
-class RandomAgent:
-    def act(self, obs, action_set, deterministic):
-        action = np.random.choice(action_set)
-        return action, None
